Updated files
I've added an email capture that's connected to an SQL database and I'm storing user email inputs and the designated prize info (awarded by spinning the wheel) to an sql database for future reference. I've also hidden the spin button so users cannot play without entering an email address. likewise, when they can play, the button is immediately hidden again after users interact with it so users cannot multi-press and impede the app logic flow. ie no cheating. I've also added an initial sound effect and have redone the gif so that it's included in the project file and doesn't lead users away from the content.